# perf: improve validate mesg sec/op delta by -72%